+++ general_utilities/job_management/subjob_test/subjob_test.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 from general_utilities.job_management.subjob_test.subjob_subpackage.tabix_subjob_testing import tabix_subjob
 from general_utilities.job_management.subjob_utility import SubjobUtility
-from general_utilities.association_resources import download_dxfile_by_name, generate_linked_dx_file
+from general_utilities.import_utils.file_handlers.dnanexus_utilities import download_dxfile_by_name, generate_linked_dx_file
 from general_utilities.mrc_logger import MRCLogger
